1. post 패키지 정리: 핵사고날 아키텍처 적용
