+++ src/main/java/com/nexters/phochak/hashtag/domain/HashtagCustomRepository.java
@@ -1,7 +1,7 @@
 package com.nexters.phochak.hashtag.domain;
 
-import com.nexters.phochak.post.PostFetchCommand;
-import com.nexters.phochak.post.PostFetchDto;
+import com.nexters.phochak.post.adapter.out.persistence.PostFetchCommand;
+import com.nexters.phochak.post.application.port.in.PostFetchDto;
 
 import java.util.List;
 import java.util.Map;
