+++ src/main/java/com/nexters/phochak/likes/application/LikeServiceImpl.java
@@ -5,10 +5,10 @@ import com.nexters.phochak.common.exception.ResCode;
 import com.nexters.phochak.likes.LikesFetchDto;
 import com.nexters.phochak.likes.domain.Likes;
 import com.nexters.phochak.likes.domain.LikesRepository;
-import com.nexters.phochak.post.PostFetchCommand;
-import com.nexters.phochak.post.PostFetchDto;
-import com.nexters.phochak.post.domain.Post;
-import com.nexters.phochak.post.domain.PostRepository;
+import com.nexters.phochak.post.adapter.out.persistence.Post;
+import com.nexters.phochak.post.adapter.out.persistence.PostFetchCommand;
+import com.nexters.phochak.post.adapter.out.persistence.PostRepository;
+import com.nexters.phochak.post.application.port.in.PostFetchDto;
 import com.nexters.phochak.user.adapter.out.persistence.UserEntity;
 import com.nexters.phochak.user.adapter.out.persistence.UserRepository;
 import lombok.RequiredArgsConstructor;
