+++ src/main/java/com/nexters/phochak/likes/application/LikesService.java
@@ -1,8 +1,8 @@
 package com.nexters.phochak.likes.application;
 
 import com.nexters.phochak.likes.LikesFetchDto;
-import com.nexters.phochak.post.PostFetchCommand;
-import com.nexters.phochak.post.PostFetchDto;
+import com.nexters.phochak.post.adapter.out.persistence.PostFetchCommand;
+import com.nexters.phochak.post.application.port.in.PostFetchDto;
 
 import java.util.List;
 import java.util.Map;
