+++ src/main/java/com/nexters/phochak/likes/domain/Likes.java
@@ -1,7 +1,7 @@
 package com.nexters.phochak.likes.domain;
 
 import com.nexters.phochak.common.domain.BaseTime;
-import com.nexters.phochak.post.domain.Post;
+import com.nexters.phochak.post.adapter.out.persistence.Post;
 import com.nexters.phochak.user.adapter.out.persistence.UserEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
