+++ src/main/java/com/nexters/phochak/post/application/port/in/CommonResponseDto.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.post;
+package com.nexters.phochak.post.application.port.in;
 
 import com.nexters.phochak.common.StatusResponse;
 import com.nexters.phochak.common.exception.ResCode;
@@ -7,24 +7,24 @@ import lombok.Setter;
 
 @Setter
 @Getter
-public class CommonResponse<T> {
+public class CommonResponseDto<T> {
     private StatusResponse status = new StatusResponse(ResCode.OK);
     private final T data;
 
-    public CommonResponse() {
+    public CommonResponseDto() {
         this.data = null;
     }
 
-    public CommonResponse(T data) {
+    public CommonResponseDto(T data) {
         this.data = data;
     }
 
-    public CommonResponse(ResCode resCode) {
+    public CommonResponseDto(ResCode resCode) {
         this();
         this.status = new StatusResponse(resCode);
     }
 
-    public CommonResponse(String resCode, String resMessage) {
+    public CommonResponseDto(String resCode, String resMessage) {
         this();
         this.status = new StatusResponse(resCode, resMessage);
     }
