+++ src/main/java/com/nexters/phochak/post/application/port/in/CustomCursorDto.java
@@ -1,9 +1,10 @@
-package com.nexters.phochak.post;
+package com.nexters.phochak.post.application.port.in;
 
 import com.nexters.phochak.common.exception.PhochakException;
 import com.nexters.phochak.common.exception.ResCode;
+import com.nexters.phochak.post.adapter.out.persistence.PostFilter;
+import com.nexters.phochak.post.adapter.out.persistence.PostSortOption;
 import com.nexters.phochak.post.domain.PostCategoryEnum;
-import com.nexters.phochak.post.domain.PostSortOption;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
@@ -12,7 +13,7 @@ import java.util.Objects;
 
 @ToString
 @Getter
-public class CustomCursor {
+public class CustomCursorDto {
     private static final int DEFAULT_PAGE_SIZE = 5;
 
     private Long lastId;
@@ -24,7 +25,7 @@ public class CustomCursor {
     private PostCategoryEnum category;
 
     @Builder
-    public CustomCursor(Long lastId, Integer pageSize, PostSortOption sortOption, Integer sortValue, PostFilter filter, Long targetUserId, PostCategoryEnum category) {
+    public CustomCursorDto(Long lastId, Integer pageSize, PostSortOption sortOption, Integer sortValue, PostFilter filter, Long targetUserId, PostCategoryEnum category) {
         this.lastId = lastId;
         this.pageSize = Objects.isNull(pageSize) ? DEFAULT_PAGE_SIZE : pageSize;
         this.sortOption = Objects.isNull(sortOption) ? PostSortOption.LATEST : sortOption;
