+++ src/main/java/com/nexters/phochak/post/application/port/in/PostFetchDto.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.post;
+package com.nexters.phochak.post.application.port.in;
 
 import com.nexters.phochak.post.domain.PostCategoryEnum;
 import com.nexters.phochak.shorts.domain.ShortsStateEnum;
