+++ src/main/java/com/nexters/phochak/report/domain/ReportPost.java
@@ -1,7 +1,7 @@
 package com.nexters.phochak.report.domain;
 
 import com.nexters.phochak.common.domain.BaseTime;
-import com.nexters.phochak.post.domain.Post;
+import com.nexters.phochak.post.adapter.out.persistence.Post;
 import com.nexters.phochak.user.adapter.out.persistence.UserEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
