+++ src/main/java/com/nexters/phochak/shorts/application/ShortsService.java
@@ -1,6 +1,6 @@
 package com.nexters.phochak.shorts.application;
 
-import com.nexters.phochak.post.domain.Post;
+import com.nexters.phochak.post.adapter.out.persistence.Post;
 import com.nexters.phochak.shorts.EncodingCallbackRequestDto;
 
 public interface ShortsService {
