+++ src/test/java/com/nexters/phochak/deprecated/integration/LikesIntegrationTest.java
@@ -7,9 +7,9 @@ import com.nexters.phochak.common.exception.ResCode;
 import com.nexters.phochak.likes.domain.Likes;
 import com.nexters.phochak.likes.domain.LikesRepository;
 import com.nexters.phochak.likes.presentation.LikesController;
-import com.nexters.phochak.post.domain.Post;
+import com.nexters.phochak.post.adapter.out.persistence.Post;
+import com.nexters.phochak.post.adapter.out.persistence.PostRepository;
 import com.nexters.phochak.post.domain.PostCategoryEnum;
-import com.nexters.phochak.post.domain.PostRepository;
 import com.nexters.phochak.shorts.domain.Shorts;
 import com.nexters.phochak.shorts.domain.ShortsRepository;
 import com.nexters.phochak.user.adapter.out.persistence.UserEntity;
