+++ src/test/java/com/nexters/phochak/deprecated/integration/PostNcpIntegrationTest.java
@@ -5,10 +5,10 @@ import com.nexters.phochak.common.docs.RestDocs;
 import com.nexters.phochak.common.exception.CustomExceptionHandler;
 import com.nexters.phochak.hashtag.domain.Hashtag;
 import com.nexters.phochak.hashtag.domain.HashtagRepository;
-import com.nexters.phochak.post.domain.Post;
+import com.nexters.phochak.post.adapter.in.web.PostController;
+import com.nexters.phochak.post.adapter.out.persistence.Post;
+import com.nexters.phochak.post.adapter.out.persistence.PostRepository;
 import com.nexters.phochak.post.domain.PostCategoryEnum;
-import com.nexters.phochak.post.domain.PostRepository;
-import com.nexters.phochak.post.presentation.PostController;
 import com.nexters.phochak.report.domain.ReportPostRepository;
 import com.nexters.phochak.report.presentation.ReportNotificationFeignClient;
 import com.nexters.phochak.shorts.domain.Shorts;
