+++ src/test/java/com/nexters/phochak/deprecated/service/LikesServiceTest.java
@@ -4,8 +4,8 @@ import com.nexters.phochak.common.exception.PhochakException;
 import com.nexters.phochak.likes.application.LikeServiceImpl;
 import com.nexters.phochak.likes.domain.Likes;
 import com.nexters.phochak.likes.domain.LikesRepository;
-import com.nexters.phochak.post.domain.Post;
-import com.nexters.phochak.post.domain.PostRepository;
+import com.nexters.phochak.post.adapter.out.persistence.Post;
+import com.nexters.phochak.post.adapter.out.persistence.PostRepository;
 import com.nexters.phochak.user.adapter.out.persistence.UserEntity;
 import com.nexters.phochak.user.adapter.out.persistence.UserRepository;
 import org.junit.jupiter.api.Disabled;
